data/index/pyproject.toml
[project]
name = "index"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb-client>=0.6.2",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.3.1",
]
